Increased Abstraction! And no notable code changes....
