+++ current.py
@@ -1,0 +1,16 @@
+import gspread
+import datetime
+from oauth2client.service_account import ServiceAccountCredentials
+
+scope = ["https://spreadsheets.google.com/feeds", 'https://www.googleapis.com/auth/spreadsheets',
+          "https://www.googleapis.com/auth/drive.file", "https://www.googleapis.com/auth/drive"]
+
+creds = ServiceAccountCredentials.from_json_keyfile_name("sei_creds.json", scope)
+
+client = gspread.authorize(creds)
+
+sheet = client.open("Inventory Backend").sheet2  # Open the Backend, look at sheet 2
+
+
+def check_location(barcode: str):
+    print("lol")
